Water temperature sensor + update intervals
Added sampling of a thermistor, so I can monitor the water temperature.
For my HomeSeer installation to be able to receive decimals, I have to
replace . with , for the reading.
The update interval is set to 5 min.

Added a visual alarm (LED) if there's no water coming from the faucet.